Fix static analysis issues
Test done: All issues have been fixed correctly.

Tracked-On: OAM-126224

+++ ta/gatekeeper/failure_record.c
@@ -135,7 +135,7 @@ uint64_t GetTimestamp(void)
 {
 	TEE_Time secure_time;
 	TEE_GetSystemTime(&secure_time);
-	return secure_time.seconds*1000 + secure_time.millis;
+	return (uint64_t)secure_time.seconds*1000 + secure_time.millis;
 }
